+++ src/lib.rs
@@ -7,6 +7,7 @@
 //! - Enable/disable the device.
 //! - Enable/disable the RGB converter.
 //! - Set RGB converter gain.
+//! - Set the number of integration cycles.
 //! - Read status of RGB converter.
 //! - Read the clear (unfiltered) channel measurement.
 //! - Read the red channel measurement.
@@ -84,6 +85,8 @@ use hal::blocking::i2c;
 pub enum Error<E> {
     /// I²C bus error
     I2C(E),
+    /// Invalid input data provided.
+    InvalidInputData
 }
 
 /// RGB converter gain
@@ -105,6 +108,7 @@ struct Register;
 
 impl Register {
     const ENABLE   : u8 = 0x00;
+    const ATIME    : u8 = 0x01;
     const CONTROL  : u8 = 0x0F;
     const ID       : u8 = 0x12;
     const STATUS   : u8 = 0x13;