+++ tests/common/mod.rs
@@ -9,6 +9,7 @@ pub struct Register;
 #[allow(unused)]
 impl Register {
     pub const ENABLE   : u8 = 0x00;
+    pub const ATIME    : u8 = 0x01;
     pub const CONTROL  : u8 = 0x0F;
     pub const ID       : u8 = 0x12;
     pub const STATUS   : u8 = 0x13;